requirements.txt
numpy
numpy-stl
matplotlib
fastapi